refactor npc combat and facing
