+++ Assets/Scripts/NPC/CombatScripts/GoblinWarChief/GoblinWarChiefCombat.cs
@@ -1,10 +1,6 @@
 using System.Collections;
 using UnityEngine;
 using Combat;
-using EquipmentSystem;
-using Skills;
-using Player;
-using Pets;
 
 namespace NPC
 {
@@ -12,18 +8,8 @@ namespace NPC
     /// Combat controller for the Goblin War Chief. Performs standard melee attacks
     /// and executes a periodic slam dealing area damage with visual effects.
     /// </summary>
-    [RequireComponent(typeof(NpcCombatant))]
-    public class GoblinWarChiefCombat : MonoBehaviour
+    public class GoblinWarChiefCombat : BaseNpcCombat
     {
-        private NpcCombatant combatant;
-        private NpcWanderer wanderer;
-        private CombatTarget currentTarget;
-        private PlayerCombatTarget playerTarget;
-        private bool hasHitPlayer;
-        private Vector2 spawnPosition;
-        private NpcSpriteAnimator spriteAnimator;
-        private SpriteRenderer spriteRenderer;
-
         [SerializeField] private float slamInterval = 10f;
         [SerializeField] private int slamDamage = 10;
         [SerializeField] private GameObject slamDustPrefab;
@@ -31,99 +17,11 @@ namespace NPC
         [SerializeField] private float shakeDuration = 0.2f;
         [SerializeField] private float shakeMagnitude = 0.1f;
 
-        private void Awake()
+        public override void BeginAttacking(CombatTarget target)
         {
-            combatant = GetComponent<NpcCombatant>();
-            wanderer = GetComponent<NpcWanderer>();
-            playerTarget = FindObjectOfType<PlayerCombatTarget>();
-            spawnPosition = transform.position;
-            spriteAnimator = GetComponent<NpcSpriteAnimator>() ?? GetComponentInChildren<NpcSpriteAnimator>();
-            spriteRenderer = GetComponent<SpriteRenderer>() ?? GetComponentInChildren<SpriteRenderer>();
-        }
-
-        private void Update()
-        {
-            var profile = combatant.Profile;
-            if (profile == null || !profile.IsAggressive)
-                return;
-
-            if (playerTarget == null)
-                playerTarget = FindObjectOfType<PlayerCombatTarget>();
-
-            if (playerTarget == null)
-                return;
-
-            float playerDistFromSpawn = Vector2.Distance(playerTarget.transform.position, spawnPosition);
-            float npcDistFromSpawn = Vector2.Distance(transform.position, spawnPosition);
-            if (currentTarget == null)
-            {
-                if (playerDistFromSpawn <= profile.AggroRange)
-                    BeginAttacking(playerTarget);
-            }
-            else if (playerDistFromSpawn > profile.AggroRange || npcDistFromSpawn > profile.AggroRange)
-            {
-                BeginAttacking(null);
-            }
-
-            if (currentTarget != null && currentTarget.IsAlive && combatant.IsAlive)
-            {
-                Vector2 diff = currentTarget.transform.position - transform.position;
-                int facingDir;
-                if (Mathf.Abs(diff.x) > Mathf.Abs(diff.y))
-                    facingDir = diff.x < 0f ? 1 : 2;
-                else
-                    facingDir = diff.y < 0f ? 0 : 3;
-
-                if (spriteAnimator != null)
-                    spriteAnimator.SetFacing(facingDir);
-                else if (spriteRenderer != null)
-                    spriteRenderer.flipX = facingDir == 2;
-            }
-        }
-
-        public void BeginAttacking(CombatTarget target)
-        {
-            if (target != null && currentTarget == target)
-                return;
-            StopAllCoroutines();
-            wanderer?.ExitCombat();
-            currentTarget = target;
-            hasHitPlayer = false;
+            base.BeginAttacking(target);
             if (target != null)
-            {
-                wanderer?.EnterCombat(target.transform);
-                StartCoroutine(AttackRoutine(target));
                 StartCoroutine(SlamRoutine(target));
-            }
-        }
-
-        private IEnumerator AttackRoutine(CombatTarget target)
-        {
-            var wait = new WaitForSeconds(4 * CombatMath.TICK_SECONDS);
-
-            // Wait until the player is within melee range before performing the first attack.
-            while (target != null && target.IsAlive && combatant.IsAlive &&
-                   Vector2.Distance(target.transform.position, transform.position) > CombatMath.MELEE_RANGE)
-            {
-                // Poll each frame until the target is close enough or combat ends.
-                yield return null;
-            }
-
-            while (target != null && target.IsAlive && combatant.IsAlive)
-            {
-                float distance = Vector2.Distance(target.transform.position, transform.position);
-                // If we move beyond aggro bounds or the target leaves melee range, stop attacking.
-                var profile = combatant.Profile;
-                float npcDistFromSpawn = Vector2.Distance(transform.position, spawnPosition);
-                if ((profile != null && npcDistFromSpawn > profile.AggroRange) || distance > CombatMath.MELEE_RANGE)
-                    break;
-
-                ResolveAttack(target);
-                yield return wait;
-            }
-
-            wanderer?.ExitCombat();
-            currentTarget = null;
         }
 
         private IEnumerator SlamRoutine(CombatTarget target)
@@ -135,68 +33,6 @@ namespace NPC
                 if (target == null || !target.IsAlive || !combatant.IsAlive)
                     break;
                 PerformSlam(target);
-            }
-        }
-
-        private void ResolveAttack(CombatTarget target)
-        {
-            var attacker = combatant.GetCombatantStats();
-            CombatantStats defender;
-            var playerTarget = target as PlayerCombatTarget;
-            if (playerTarget != null)
-            {
-                var skills = playerTarget.GetComponent<SkillManager>();
-                var equipment = playerTarget.GetComponent<EquipmentAggregator>();
-                var loadout = playerTarget.GetComponent<PlayerCombatLoadout>();
-                defender = CombatantStats.ForPlayer(skills, equipment,
-                    loadout != null ? loadout.Style : CombatStyle.Defensive,
-                    DamageType.Melee);
-            }
-            else
-            {
-                defender = new CombatantStats
-                {
-                    AttackLevel = 1,
-                    StrengthLevel = 1,
-                    DefenceLevel = 1,
-                    Equip = new EquipmentAggregator.CombinedStats(),
-                    Style = CombatStyle.Defensive,
-                    DamageType = target.PreferredDefenceType
-                };
-            }
-
-            int attEff = CombatMath.GetEffectiveAttack(attacker.AttackLevel, attacker.Style);
-            int defEff = CombatMath.GetEffectiveDefence(defender.DefenceLevel, defender.Style);
-            int atkRoll = CombatMath.GetAttackRoll(attEff, attacker.Equip.attack);
-            int defBonus = defender.DamageType switch
-            {
-                DamageType.Magic => defender.Equip.magicDef,
-                DamageType.Ranged => defender.Equip.rangeDef,
-                _ => defender.Equip.meleeDef
-            };
-            int defRoll = CombatMath.GetDefenceRoll(defEff, defBonus);
-            bool hit = Random.value < CombatMath.ChanceToHit(atkRoll, defRoll);
-            int damage = 0;
-            if (hit)
-            {
-                int strEff = CombatMath.GetEffectiveStrength(attacker.StrengthLevel, attacker.Style);
-                int maxHit = CombatMath.GetMaxHit(strEff, attacker.Equip.strength);
-                damage = CombatMath.RollDamage(maxHit);
-                target.ApplyDamage(damage, attacker.DamageType, this);
-                var targetName = (target as MonoBehaviour)?.name ?? "target";
-                Debug.Log($"{name} dealt {damage} damage to {targetName}.");
-            }
-            else
-            {
-                var targetName = (target as MonoBehaviour)?.name ?? "target";
-                Debug.Log($"{name} missed {targetName}.");
-            }
-
-            if (!hasHitPlayer && playerTarget != null && PetDropSystem.GuardModeEnabled)
-            {
-                var pet = PetDropSystem.ActivePetCombat;
-                pet?.CommandAttack(combatant);
-                hasHitPlayer = true;
             }
         }
